+++ changelog.md
@@ -5,6 +5,7 @@
 - Bug fix straneg behavior on the hook render actor need a mrge object for set up the correct values of the flags
 - Integration with [DragonFlagon Quality of Life, Vehicle Cargo Capacity Unit Feature](https://github.com/flamewave000/dragonflagon-fvtt/blob/master/df-qol/README.md#vehicle-cargo-capacity-unit)
 - Add new settings 'useStrengthMultiplier'
+- Add new settings 'useVarianEncumbranceWithSpecificType'
 - Many big fix
 
 ### Update 0.6.8